pyproject.toml
[project]
name = "rbac-template"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "passlib[argon2]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pycryptodome>=3.23.0",
    "pyjwt>=2.10.1",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "redis>=6.3.0",
    "setuptools==80",
    "tortoise-orm[asyncpg]>=0.25.1",
    "uvicorn[standard]>=0.35.0",
]
